backend/requirements.txt
Flask
gunicorn
SQLAlchemy
psycopg2-binary
pandas
prophet
requests
python-dotenv
geoalchemy2
Shapely
python-dateutil
